+++ docs/app/Examples/modules/Progress/ProgressExamples.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react'
 import Types from './Types/Types'
 import Content from './Content/Content'
 import States from './States/States'
+import Variations from './Variations/Variations'
 
 export default class ProgressExamples extends Component {
   render() {
@@ -10,6 +11,7 @@ export default class ProgressExamples extends Component {
         <Types />
         <Content />
         <States />
+        <Variations />
       </div>
     )
   }