+++ docs/app/Examples/modules/Progress/States/Active.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Progress } from 'stardust'
 
 const ProgressActiveExample = () => (
-  <Progress percent={60} className='active'>
+  <Progress percent={60} active>
     Active
   </Progress>
 )