+++ src/index.js
@@ -1,3 +1,4 @@
+import './utils/jquery'
 import { deprecateComponent } from './utils/deprecate'
 
 // ----------------------------------------
@@ -45,6 +46,7 @@ export Icon from './elements/Icon/Icon'
 export Image from './elements/Image/Image'
 export Input from './elements/Input/Input'
 export Label from './elements/Label/Label'
+export Loader from './elements/Loader/Loader'
 
 import _List from './elements/List/List'
 export { _List as List }
@@ -52,6 +54,7 @@ export const ListItem = deprecateComponent('ListItem', 'Use "List.Item" instead.
 
 export Segment from './elements/Segment/Segment'
 export Segments from './elements/Segment/SegmentSegments'
+export Rail from './elements/Rail/Rail'
 
 // ----------------------------------------
 // Modules